+++ bugsnag-android-performance-appcompat/src/test/kotlin/com/bugsnag/android/performance/FragmentActivityLifecycleCallbacksTest.kt
@@ -65,6 +65,7 @@ class FragmentActivityLifecycleCallbacksTest {
     @Test
     @Suppress("DestructuringDeclarationWithTooManyEntries")
     fun testInterleavedLifecycles() {
+        FragmentInstrumentation.enabled = true
         lifecycleCallbacks.onFragmentPreCreated(fm, fragment1, null)
         assertNotEquals(
             "Fragment.onCreate should have a valid SpanContext",
@@ -118,6 +119,18 @@ class FragmentActivityLifecycleCallbacksTest {
         lifecycleCallbacks.onFragmentStarted(fm, doNotAutoInstrumentFragment)
         lifecycleCallbacks.onFragmentResumed(fm, doNotAutoInstrumentFragment)
 
+        assertEquals(0, spanCollector.size)
+    }
+
+    @Test
+    fun testDisableFragmentInstrument() {
+        FragmentInstrumentation.enabled = false
+        lifecycleCallbacks.onFragmentPreCreated(fm, fragment1, null)
+
+        assertEquals(0, spanCollector.size)
+
+        lifecycleCallbacks.onFragmentStarted(fm, fragment1)
+        lifecycleCallbacks.onFragmentResumed(fm, fragment1)
         assertEquals(0, spanCollector.size)
     }
 }