+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 2.2.0 (2026-01-12)
+
+### Changes
+
+* System metrics are now reported for first-class `ViewLoad` spans
+  [#475](https://github.com/bugsnag/bugsnag-android-performance/pull/475)
+* Enable runtime control of fragment instrumentation via a new `FragmentInstrumentation.enabled` flag
+  [#490](https://github.com/bugsnag/bugsnag-android-performance/pull/490)
+
+### Bug fixes
+
+* Fixed an issue where `onSpanEnd` callbacks were not invoked for sampled spans
+  [#486](https://github.com/bugsnag/bugsnag-android-performance/pull/486)
+
 ## 2.1.1 (2025-11-20)
 
 ### Bug fixes